古い部屋を自動削除できない不具合修正

+++ app/models/room_member.rb
@@ -9,6 +9,7 @@ class RoomMember < ApplicationRecord
   end
 
   def before_room_member_destroy
+    return unless self.player
     self.player.destroy if self.player.is_guest_user
   end
